// src/types/relationsPersonne.ts
export interface TypeRelation {
  id: string
  code: string | null
  id_inverse_defaut: string | null
  id_inverse_masculin: string | null
  id_inverse_feminin: string | null
  id_traduction: string | null
  id_langue: string | null
  code_langue: string | null
  nom_langue: string | null
  libelle: string | null
}